Move CounterList out of tests
